add Phong shadows

+++ src/materials.h
@@ -6,7 +6,7 @@
     double k_a;//ambiant component, between 0 and 1
     double k_d;//diffuse component, between 0 and 1
     double k_s;//specular component, between 0 and 1
-    double alpha;//material shine, more than 1. The larger it is, the shinest the surface is.
+    int alpha;//material shine, more than 1. The larger it is, the shinest the surface is.
   };
 
   Material MATERIAL_create_material(double k_a, double k_d, double k_s, double alpha);
